refactor(collector): API クライアントを関数分解し重複チェックをページ単位に移動
JobStoreClient（Effect.Service）を解体し、lib/apiClient/ 配下に query.ts
（filterUnregistered）と mutation.ts（insertJob/upsertCompany）として独立
関数で再編成した。環境変数は APIConfig（Context.Tag）に集約。
JobDetailLoader はインターフェースとして残し、mutation 関数のアダプタに。

重複チェック（filterUnregistered）は crawl の最終段ではなくページ単位
で呼ぶように変更し、ページ遷移と 2 秒のレート制限待機を pipe で束ねた
（最終ページでは待たない）。クロール結果は JobNumber[] にフラット化し、
SQS 送信側で {jobNumber} を再構築。

ES2025 の Set.prototype.difference を使うため tsconfig を ESNext に。
