+++ apps/backend/collector/tsconfig.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
-    "target": "ES2022",
+    "target": "ESNext",
     "module": "ESNext",
     "moduleResolution": "Bundler",
-    "lib": ["es2022", "DOM"],
+    "lib": ["ESNext", "DOM"],
     "strict": true,
     "noImplicitAny": true,
     "strictNullChecks": true,
